backend/requirements.txt
# MeetLove Backend

fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-socketio>=4.7.0
aiohttp>=3.8.0
opencv-python>=4.8.0
numpy>=1.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6